Rename directory and view model
